tests/wpt/tests/scroll-animations/animation-trigger/parsing/timeline-trigger-behavior-computed.tentative.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Animations: getComputedStyle().animationTriggerBehavior</title>
<link rel="help" href="https://drafts.csswg.org/css-animations-2/#animation-trigger">
<meta name="assert" content="timeline-trigger-type computed value is as specified.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value("timeline-trigger-behavior", "initial", "once");
test_computed_value("timeline-trigger-behavior", "once, repeat, alternate, state");
</script>
</body>
</html>
